feat: Update GeneralInfoDto to use double for totalSpent and adjust related queries

+++ order_service/src/main/java/com/lamnguyen/order_service/domain/dto/GeneralInfoDto.java
@@ -8,9 +8,9 @@ import lombok.Setter;
 public  class GeneralInfoDto {
 	long userId;
 	long totalOrder;
-	long totalSpent;
+	double totalSpent;
 
-	public GeneralInfoDto(long userId, long totalOrder, long totalSpent) {
+	public GeneralInfoDto(long userId, long totalOrder, double totalSpent) {
 		this.userId = userId;
 		this.totalOrder = totalOrder;
 		this.totalSpent = totalSpent;
